Fix histograms

+++ common/telemetry/metrics/instruments/SchedulerInstruments.hpp
@@ -23,6 +23,6 @@ namespace cta::telemetry::metrics {
 
 extern std::unique_ptr<opentelemetry::metrics::Counter<uint64_t>> schedulerQueueingCounter;
 extern std::unique_ptr<opentelemetry::metrics::Counter<uint64_t>> objectstoreLockAcquireCounter;
-extern std::unique_ptr<opentelemetry::metrics::Histogram<double>> objectstoreLockAcquireDurationHistogram;
+extern std::unique_ptr<opentelemetry::metrics::Histogram<uint64_t>> objectstoreLockAcquireDurationHistogram;
 
 }  // namespace cta::telemetry::metrics
